+++ framework-crates/objc2-foundation/src/tests/auto_traits.rs
@@ -1,12 +1,11 @@
-#![cfg(feature = "all")] // TODO: More precise
+#![allow(dead_code, unused_imports)]
 use core::panic::{RefUnwindSafe, UnwindSafe};
 
 use static_assertions::assert_not_impl_any;
 
-use crate::*;
 use objc2::define_class;
 use objc2::rc::Retained;
-use objc2::runtime::AnyObject;
+use objc2::runtime::{AnyObject, NSObject};
 
 // We expect most Foundation types to be UnwindSafe and RefUnwindSafe,
 // since they follow Rust's usual mutability rules (&T = immutable).
@@ -45,49 +44,83 @@ fn test_generic_auto_traits() {
     // assert_unwindsafe::<Retained<NSDictionary<NSProcessInfo, NSProcessInfo>>>();
 
     // TODO: Unpin?
-    assert_not_impl_any!(NSArray<AnyObject>: Unpin);
-    assert_not_impl_any!(NSMutableArray<AnyObject>: Unpin);
-    assert_not_impl_any!(NSDictionary<AnyObject, AnyObject>: Unpin);
+    #[cfg(feature = "NSArray")]
+    assert_not_impl_any!(crate::NSArray<AnyObject>: Unpin);
+    #[cfg(feature = "NSArray")]
+    assert_not_impl_any!(crate::NSMutableArray<AnyObject>: Unpin);
+    #[cfg(feature = "NSDictionary")]
+    assert_not_impl_any!(crate::NSDictionary<AnyObject, AnyObject>: Unpin);
 
     // Collections are not Send + Sync, since they are interior mutable, i.e.
     // mutable from `&self`.
-    assert_not_impl_any!(NSArray<SendSyncObject>: Send, Sync);
-    assert_not_impl_any!(NSMutableArray<SendSyncObject>: Send, Sync);
-    assert_not_impl_any!(NSDictionary<SendSyncObject, SendSyncObject>: Send, Sync);
+    #[cfg(feature = "NSArray")]
+    assert_not_impl_any!(crate::NSArray<SendSyncObject>: Send, Sync);
+    #[cfg(feature = "NSArray")]
+    assert_not_impl_any!(crate::NSMutableArray<SendSyncObject>: Send, Sync);
+    #[cfg(feature = "NSDictionary")]
+    assert_not_impl_any!(crate::NSDictionary<SendSyncObject, SendSyncObject>: Send, Sync);
 
     // TODO: Make these UnwindSafe?
-    assert_not_impl_any!(NSDictionary<NSProcessInfo, NSProcessInfo>: UnwindSafe, RefUnwindSafe);
-    assert_not_impl_any!(NSSet<NSProcessInfo>: UnwindSafe, RefUnwindSafe);
-    assert_not_impl_any!(Retained<NSSet<NSProcessInfo>>: UnwindSafe, RefUnwindSafe);
-    assert_not_impl_any!(NSMutableArray<NSProcessInfo>: UnwindSafe, RefUnwindSafe);
-    assert_not_impl_any!(NSMutableDictionary<NSProcessInfo, NSProcessInfo>: UnwindSafe, RefUnwindSafe);
-    assert_not_impl_any!(NSMutableSet<NSProcessInfo>: UnwindSafe, RefUnwindSafe);
+    #[cfg(feature = "NSProcessInfo")]
+    {
+        use crate::NSProcessInfo;
+
+        #[cfg(feature = "NSDictionary")]
+        assert_not_impl_any!(crate::NSDictionary<NSProcessInfo, NSProcessInfo>: UnwindSafe, RefUnwindSafe);
+        #[cfg(feature = "NSSet")]
+        assert_not_impl_any!(crate::NSSet<NSProcessInfo>: UnwindSafe, RefUnwindSafe);
+        #[cfg(feature = "NSSet")]
+        assert_not_impl_any!(Retained<crate::NSSet<NSProcessInfo>>: UnwindSafe, RefUnwindSafe);
+        #[cfg(feature = "NSArray")]
+        assert_not_impl_any!(crate::NSMutableArray<NSProcessInfo>: UnwindSafe, RefUnwindSafe);
+        #[cfg(feature = "NSDictionary")]
+        assert_not_impl_any!(crate::NSMutableDictionary<NSProcessInfo, NSProcessInfo>: UnwindSafe, RefUnwindSafe);
+        #[cfg(feature = "NSSet")]
+        assert_not_impl_any!(crate::NSMutableSet<NSProcessInfo>: UnwindSafe, RefUnwindSafe);
+    }
 }
 
 #[test]
 fn send_sync_unwindsafe() {
-    assert_unwindsafe::<NSAttributedString>();
-    assert_auto_traits::<NSComparisonResult>();
-    assert_unwindsafe::<NSData>();
+    #[cfg(feature = "NSAttributedString")]
+    assert_unwindsafe::<crate::NSAttributedString>();
+    #[cfg(feature = "NSObjCRuntime")]
+    assert_auto_traits::<crate::NSComparisonResult>();
+    #[cfg(feature = "NSData")]
+    assert_unwindsafe::<crate::NSData>();
     // TODO: Figure out if Send + Sync is safe?
-    // assert_auto_traits::<NSEnumerator2<NSProcessInfo>>();
-    // assert_auto_traits::<NSFastEnumerator2<NSArray<NSProcessInfo>>>();
-    assert_auto_traits::<NSError>();
-    assert_auto_traits::<NSException>();
-    assert_auto_traits::<NSPoint>();
-    assert_auto_traits::<NSRect>();
-    assert_auto_traits::<NSSize>();
-    assert_unwindsafe::<NSMutableAttributedString>();
-    assert_unwindsafe::<NSMutableData>();
-    assert_unwindsafe::<NSMutableString>();
-    assert_auto_traits::<NSNumber>();
-    // assert_auto_traits::<NSObject>(); // Intentional
-    assert_auto_traits::<NSProcessInfo>();
-    assert_auto_traits::<NSRange>();
-    assert_unwindsafe::<NSString>();
-    assert_unwindsafe::<MainThreadMarker>(); // Intentional
-    assert_auto_traits::<NSThread>();
-    assert_auto_traits::<NSUUID>();
-    // assert_auto_traits::<NSValue>(); // Intentional
-    assert_unwindsafe::<NSZone>(); // Intentional
+    // assert_auto_traits::<crate::NSEnumerator2<crate::NSProcessInfo>>();
+    // assert_auto_traits::<crate::NSFastEnumerator2<crate::NSArray<crate::NSProcessInfo>>>();
+    #[cfg(feature = "NSError")]
+    assert_auto_traits::<crate::NSError>();
+    #[cfg(feature = "NSException")]
+    assert_auto_traits::<crate::NSException>();
+    #[cfg(all(feature = "NSGeometry", feature = "objc2-core-foundation"))]
+    assert_auto_traits::<crate::NSPoint>();
+    #[cfg(all(feature = "NSGeometry", feature = "objc2-core-foundation"))]
+    assert_auto_traits::<crate::NSRect>();
+    #[cfg(all(feature = "NSGeometry", feature = "objc2-core-foundation"))]
+    assert_auto_traits::<crate::NSSize>();
+    #[cfg(feature = "NSAttributedString")]
+    assert_unwindsafe::<crate::NSMutableAttributedString>();
+    #[cfg(feature = "NSData")]
+    assert_unwindsafe::<crate::NSMutableData>();
+    #[cfg(feature = "NSString")]
+    assert_unwindsafe::<crate::NSMutableString>();
+    #[cfg(feature = "NSValue")]
+    assert_auto_traits::<crate::NSNumber>();
+    // assert_auto_traits::<crate::NSObject>(); // Intentional
+    #[cfg(feature = "NSProcessInfo")]
+    assert_auto_traits::<crate::NSProcessInfo>();
+    #[cfg(feature = "NSRange")]
+    assert_auto_traits::<crate::NSRange>();
+    #[cfg(feature = "NSString")]
+    assert_unwindsafe::<crate::NSString>();
+    #[cfg(feature = "NSThread")]
+    assert_auto_traits::<crate::NSThread>();
+    #[cfg(feature = "NSUUID")]
+    assert_auto_traits::<crate::NSUUID>();
+    // assert_auto_traits::<crate::NSValue>(); // Intentional
+    #[cfg(feature = "NSZone")]
+    assert_unwindsafe::<crate::NSZone>(); // Intentional
 }